requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.0
drf-spectacular==0.28.0
